refactor(graduate): all management 경로 변경 및 pages 레이어로 분리
