--- apps/graduate/src/features/all-management/components/index.ts
@@ -1,1 +1,0 @@
-export { default as AllManagement } from './AllManagement';
