+++ apps/graduate/src/pages/admin/all/index.ts
@@ -1,0 +1,1 @@
+export { default as AllManagementPage } from './ui/AllManagementPage';
