+++ apps/graduate/src/pages/admin/all/styles/AllManagementPage.css.ts
