+++ apps/graduate/src/pages/admin/all/types/allManagement.ts
