+++ apps/graduate/src/pages/admin/all/ui/UserDetailModal/UserApplicationView.tsx
