+++ apps/graduate/src/pages/admin/all/ui/UserDetailModal/UserDetailModal.tsx
