+++ apps/graduate/src/pages/admin/all/ui/UserDetailModal/UserDetailView.tsx
