+++ apps/graduate/src/pages/admin/all/ui/UserDetailModal/UserMiddleReportView.tsx
