+++ apps/graduate/src/pages/admin/all/ui/UserDetailModal/index.ts
