+++ apps/graduate/src/routes/_afterLogin/all.tsx
@@ -1,0 +1,7 @@
+import { createFileRoute } from '@tanstack/react-router';
+
+import { AllManagementPage } from '~/pages/admin/all';
+
+export const Route = createFileRoute('/_afterLogin/all')({
+  component: AllManagementPage,
+});
