--- apps/graduate/src/routes/_afterLogin/graduates-all.tsx
@@ -1,7 +1,0 @@
-import { createFileRoute } from '@tanstack/react-router';
-
-import { AllManagement } from '~/features/all-management/components';
-
-export const Route = createFileRoute('/_afterLogin/graduates-all')({
-  component: AllManagement,
-});
